Added some nice features.
Now has the verbose and interactive options. These tell the program to
print extra information and ask for extra confirmation respectively.
Also now correctly expands paths in the config file and changes the
working directory of the program to the directory specified by the
program.

Also added a util file for a couple functions that I fould that I
wanted but couldn't find anywhere to put.

+++ src/removeaction.cc
@@ -24,6 +24,8 @@
 
 #include <err.h>
 
+#include "util.h"
+
 namespace dfm {
 
 RemoveAction::RemoveAction(const boost::filesystem::path& filePath)
@@ -52,6 +54,11 @@ RemoveAction::setFilePath(const boost::filesystem::path& filePath)
 bool
 RemoveAction::performAction()
 {
+    if (isInteractive()) {
+        std::string prompt = "Remove file " + filePath.string() + "?";
+        if (!getYesOrNo(prompt))
+            return true;
+    }
     try {
         if (boost::filesystem::exists(filePath)) {
             boost::filesystem::remove_all(filePath);
